blockstorage: Move CBlockFileInfo to blocktreestorage
This eliminates the circular dependency between the blockstorage and the
blocktreestorage modules again.
