+++ src/kernel/blocktreestorage.h
@@ -7,7 +7,6 @@
 
 #include <kernel/cs_main.h>
 #include <serialize.h>
-#include <streams.h>
 #include <sync.h>
 #include <util/fs.h>
 
@@ -30,8 +29,6 @@ class SignalInterrupt;
 }
 
 namespace kernel {
-
-class CBlockFileInfo;
 
 // Checksums are calculated from the serialized value and its position in the
 // file. This protects against out of order data and allows the same checksum
@@ -93,6 +90,50 @@ public:
 
     WriterLock(const WriterLock&) = delete;
     WriterLock& operator=(const WriterLock&) = delete;
+};
+
+class CBlockFileInfo
+{
+public:
+    uint32_t nBlocks{};      //!< number of blocks stored in file
+    uint32_t nSize{};        //!< number of used bytes of block file
+    uint32_t nUndoSize{};    //!< number of used bytes in the undo file
+    uint32_t nHeightFirst{}; //!< lowest height of block in file
+    uint32_t nHeightLast{};  //!< highest height of block in file
+    uint64_t nTimeFirst{};   //!< earliest time of block in file
+    uint64_t nTimeLast{};    //!< latest time of block in file
+
+    // Note: The SERIALIZE_METHODS here use VARINT encoding for compatibility with
+    // the legacy leveldb block tree db, used during migration in CreateAndMigrateBlockTree.
+    // BlockFileInfoWrapper uses fixed-width encoding for the new flat file storage.
+    SERIALIZE_METHODS(CBlockFileInfo, obj)
+    {
+        READWRITE(VARINT(obj.nBlocks));
+        READWRITE(VARINT(obj.nSize));
+        READWRITE(VARINT(obj.nUndoSize));
+        READWRITE(VARINT(obj.nHeightFirst));
+        READWRITE(VARINT(obj.nHeightLast));
+        READWRITE(VARINT(obj.nTimeFirst));
+        READWRITE(VARINT(obj.nTimeLast));
+    }
+
+    CBlockFileInfo() = default;
+
+    std::string ToString() const;
+
+    /** update statistics (does not update nSize) */
+    void AddBlock(unsigned int nHeightIn, uint64_t nTimeIn)
+    {
+        if (nBlocks == 0 || nHeightFirst > nHeightIn)
+            nHeightFirst = nHeightIn;
+        if (nBlocks == 0 || nTimeFirst > nTimeIn)
+            nTimeFirst = nTimeIn;
+        nBlocks++;
+        if (nHeightIn > nHeightLast)
+            nHeightLast = nHeightIn;
+        if (nTimeIn > nTimeLast)
+            nTimeLast = nTimeIn;
+    }
 };
 
 class BlockTreeStore
